requirements.txt
fastapi
uvicorn
pydantic
crewai
python-dotenv
requests
langchain
langchain-community
crewai-tools
firebase-admin
schedule
